Converted ViewModelBase to abstract class.

+++ OfficeApp/OfficeApp/ViewModels/EditDeleteDepartmentPageViewModel.cs
@@ -3,8 +3,10 @@ using OfficeApp.Helpers;
 using OfficeApp.Models;
 using Prism.Commands;
 using Prism.Navigation;
+using Prism.Services;
 using System.Net.Http;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace OfficeApp.ViewModels
 {
@@ -24,36 +26,47 @@ namespace OfficeApp.ViewModels
             }
         }
 
-        public EditDeleteDepartmentPageViewModel(INavigationService navigationService) : base(navigationService)
+        public EditDeleteDepartmentPageViewModel(INavigationService navigationService, IPageDialogService pageDialogService)
+            : base(navigationService, pageDialogService)
         {
         }
 
         public override void OnNavigatingTo(NavigationParameters parameters)
         {
             if (parameters.ContainsKey("(^_^)ImTheKey"))
-            {
                 CurrentDepartment = (Department)parameters["(^_^)ImTheKey"];
-            }
 
             base.OnNavigatingTo(parameters);
         }
 
-        public DelegateCommand UpdateCommand => new DelegateCommand(Update);
+        public DelegateCommand UpdateCommand => new DelegateCommand(async () => await Update());
 
-        private async void Update()
+        private async Task Update()
         {
             _client.DefaultRequestHeaders.Add("Authorization", $"Bearer {Settings.Jwt}");
 
             var content = JsonConvert.SerializeObject(CurrentDepartment);
-            await _client.PutAsync(Constants.URLs.SetDepartmentUrl() + CurrentDepartment.Id, new StringContent(content, Encoding.UTF8, "application/json"));
 
-            await NavigationService.GoBackAsync();
+            using (var response = await _client.PutAsync(Constants.URLs.SetDepartmentUrl()+ CurrentDepartment.Id,
+                                          new StringContent(content, Encoding.UTF8, "application/json")))
+            {
+                if (response.IsSuccessStatusCode)
+                {
+                    await NavigationService.GoBackAsync();
+                    return;
+                }
+
+                await PageDialogService.DisplayAlertAsync("Error updating", "Please check your internet", "OK");
+            }
         }
 
-        public DelegateCommand DeleteCommand => new DelegateCommand(Delete);
+        public DelegateCommand DeleteCommand => new DelegateCommand(async () => await Delete());
 
-        private async void Delete()
+        private async Task Delete()
         {
+            var userResponse = await PageDialogService.DisplayAlertAsync("Deleting an entry", "You sure you want to delete this?", "Yes", "Cancel");
+            if (!userResponse) return;
+
             _client.DefaultRequestHeaders.Add("Authorization", $"Bearer {Settings.Jwt}");
             await _client.DeleteAsync(Constants.URLs.SetDepartmentUrl() + CurrentDepartment.Id);
